+++ cli.js
@@ -59,6 +59,7 @@ program
     .option('--lib [bool]', 'Scaffold a new project with support for library builds')
     .option('--lazyModule [bool]', 'Scaffold a new project with support for lazyloaded modules')
     .option('--dynamicRoutes [bool]', 'Scaffold a new project with support for routes configured by JSON prior to Bootstrap')
+    .option('--electron [bool]', 'Scaffold a new project with additional files needed to support Electron, serve with Electron')
     .option('--angularVersion [string]', 'Scaffold a new project with a specific @angular version')
     .option('update, --update [bool]', 'Update a project')
     .option('--cliVersion [string]', 'Update an existing project with changes to config files provided by the cli')
@@ -80,6 +81,10 @@ if (program.scaffold) {
 
     if (program.dynamicRoutes) {
         cliCommand += ' dynamicRoutes=true';
+    }
+
+    if (program.electron) {
+        cliCommand += ' electron=true';
     }
 
     if (program.lazyModule) {
@@ -196,6 +201,13 @@ if (program.build) {
     }
     else {
         cliCommand += ' serve=false';
+    }
+
+    if (program.electron === true) {
+        cliCommand += ' electron=true';
+    }
+    else {
+        cliCommand += ' electron=false';
     }
 
     if (program.deploy === true) {